drivers:at86rf231:change at68rf231_get_channel to be transceiver API compliant

+++ drivers/at86rf231/at86rf231.c
@@ -157,7 +157,12 @@ uint8_t at86rf231_set_channel(uint8_t channel)
     return radio_channel;
 }
 
-uint8_t at86rf231_get_channel(void)
+/*
+ * @brief Get the radio channel of the device
+ *
+ * @return The current radio channel of the device.
+ */
+int32_t at86rf231_get_channel(void)
 {
     return radio_channel;
 }
